requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
pillow>=10.0.0
pyyaml>=6.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
scipy>=1.11.0
streamlit-webrtc>=0.47.0
instagrapi>=2.0.0
